% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/int_check_ties.R
\name{check_ties}
\alias{check_ties}
\title{Check Ties.}
\usage{
check_ties(first, last)
}
\arguments{
\item{first}{Predictions when setting ties.method = "first"}

\item{last}{Predictions when setting ties.method = "last"}
}
\value{
Nothing.
}
\description{
Internal function called by \code{score_lundtax}.
Not meant for out of package use.
}
\details{
Checks score ties in the prediction scores and prints a message indicating
the sample where the tie occurred, the two subtypes with the tied scores and
the subtype that is reported in the output object.
}
