% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{classifier_grade3}
\alias{classifier_grade3}
\title{Classifier Grade 3.}
\format{
A list of 6.
}
\usage{
data(classifier_grade3)
}
\description{
Classifier generated with \code{\link[multiclassPairs:predict_RF]{multiclassPairs::predict_RF()}}, this dataset is needed for predicting
the grade (1, 2, or 3) and calculating scores for incoming data.
}
\details{
A Large rule_based_RandomForest object. A list of 6.

\itemize{
\item genes. Genes in ensembl format.
\item rules. A set of rules for the classifier in ensembl format.
\item TrainingMatrix. Binary matrix for the rules in the training data.
\item boruta. Boruta results for the classifier.
\item RF_classifier. Random forest classifier details.
\item calls. Information on how the model was generated.
}
}
\keyword{datasets}
