% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{classifier_lundtax_7c}
\alias{classifier_lundtax_7c}
\title{Classifier LundTax 7c.}
\format{
A list of 6.
}
\usage{
data(classifier_lundtax_7c)
}
\description{
Classifier as a rule_based_RandomForest object. Predicts samples as one of the 3 Uro subclasses,
UroA, UroB, or UroC. Object includes the final RF classifier, the used genes and rules in the
final model, the Boruta results, and the training matrix. The training matrix is a binary matrix
containing the rule values for the training data and it is used for imputation purposes during
the prediction if values are missing in the sample. This object was generated using the
\code{\link[multiclassPairs:predict_RF]{multiclassPairs::predict_RF()}} function.
}
\details{
A Large rule_based_RandomForest object. A list of 6.

\itemize{
\item genes. Genes in hgnc format.
\item rules. A set of rules for the classifier in hgnc format.
\item TrainingMatrix. Binary matrix for the rules in the training data.
\item boruta. Boruta results for the classifier.
\item RF_classifier. Random forest classifier details.
\item calls. Information on how the model was generated.
}
}
\keyword{datasets}
