% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{cols}
\alias{cols}
\title{Colors used for Lund Taxonomy subtypes
List including hexadecimal color codes for LundTax subtypes, stage and datasets}
\format{
An object of class \code{list} of length 4.
}
\usage{
cols
}
\description{
Colors used for Lund Taxonomy subtypes
List including hexadecimal color codes for LundTax subtypes, stage and datasets
}
\keyword{datasets}
