% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{gene_info_lund}
\alias{gene_info_lund}
\title{Gene IDs for classification, scoring and plotting}
\format{
An object of class \code{data.frame} with 1900 rows and 2 columns.
}
\usage{
gene_info_lund
}
\description{
Gene IDs for classification, scoring and plotting
}
\keyword{datasets}
