% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gene_list}
\alias{gene_list}
\title{Gene List.}
\format{
A data frame with 1900 rows (genes) and 2 columns (hgnc_symbol and ensembl_gene_id).
}
\usage{
data(gene_list)
}
\description{
Gene annotations in hgnc and ensembl format. Features, for the classifier relevant genes.
For convenience, both formats are available for seamless conversion of IDs, regardless of the
format of the incoming data.
}
\details{
A data frame with gene information in different formats.

\itemize{
\item ensembl_gene_id. Gene annotation in ensembl format.
\item hgnc_symbol. Gene annotation in HUGO format.
}
}
\keyword{datasets}
