% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_data}
\alias{get_data}
\title{Get Data}
\usage{
get_data(
  this_metadata = NULL,
  these_sample_ids = NULL,
  this_data = NULL,
  return_type = "metadata",
  rows_to_column = FALSE,
  sample_column = NULL,
  exclude_id = NULL,
  first_variable = NULL,
  first_value = NULL,
  second_variable = NULL,
  second_value = NULL,
  third_variable = NULL,
  third_value = NULL
)
}
\arguments{
\item{this_metadata}{Required parameter, should be a data frame with metadata information}

\item{these_sample_ids}{Optional parameter. If provided, the incoming metadata will be restricted
to the sample IDs provided here. Should be a vector of sample IDs of interest.}

\item{this_data}{Required parameter if return_type is set to \code{expression_data}. Should be a data
frame with sample IDs as column names.}

\item{return_type}{Should be on of the following; metadata (default), sample_id or expression_data. If sample IDs
are selected, a vector with selected sample IDs will be returned, nothing else. If expression_data
is selected, the user needs to specify the incoming expression data with \code{this_data}.}

\item{rows_to_column}{Boolean parameter. Set to TRUE to convert row names to a new column called
sample_id. Default is FALSE.}

\item{sample_column}{Optional, lets the user define a specific column in the data frame with sample IDs.}

\item{exclude_id}{If metadata is provided and no sample IDs are specified, the user can control
if any sample IDs should be omitted from the return. This parameter expects a character of vectors.}

\item{first_variable}{Filtering options, should be the name of the column of interest.}

\item{first_value}{Filtering options, the filter value for the first variable.}

\item{second_variable}{Filtering options, should be the name of the column of interest.}

\item{second_value}{Filtering options, the filter value for the second variable.}

\item{third_variable}{Filtering options, should be the name of the column of interest.}

\item{thid_value}{Filtering options, the filter value for the third variable.}
}
\value{
A data frame with subset metadata, or a vector of characters with sample IDs.
}
\description{
subset and filter incoming data based on a set of criteria.
}
\details{
Takes a metadata data frame as the main input and allows easy subset for various variables.
For more information, see parameter descriptions and examples.
}
\examples{
#example 1
#return all sample IDs from bundled metadata
get_data(this_metadata = sjodahl_2017_meta, 
         return_type = "sample_id")

#use more filter options and return metadata data frame
get_data(this_metadata = sjodahl_2017_meta, 
         first_variable = "gender", 
         first_value = "Female", 
         second_variable = "turb_stage", 
         second_value = 1, 
         third_variable = "adj_chemo", 
         third_value = 0)

}
