% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_glm_scores.R
\name{get_glm_scores}
\alias{get_glm_scores}
\title{Get General Linear Model Scores.}
\usage{
get_glm_scores(
  these_predictions = NULL,
  these_samples_metadata = NULL,
  subtype_class = "5class",
  this_subtype = NULL,
  cat_variable = NULL,
  row_to_col = FALSE,
  sample_id_col = NULL
)
}
\arguments{
\item{these_predictions}{Required parameter. A data frame with predictions scores, subtype and
metadata information.}

\item{these_samples_metadata}{Required, a data frame with metadata associated with the prediction
calls, Note, the function will subset the return to samples are included in this object.}

\item{subtype_class}{Can be one of the following; 5_class or 7_class. Default is 5class.}

\item{this_subtype}{Optional parameter. Allows the user to subset the return to a specific subtype
within the selected class. If not specified, the function will return a data frame with subtype
information for all the subtypes within the specified class.}

\item{cat_variable}{Required parameter. Specify the two level categorical variable you want to test for.}

\item{row_to_col}{Boolean parameter. Set to TRUE to transform row names of the metadata to a new
column called sample_id. Default is FALSE.}

\item{sample_id_col}{Parameter dictating the column name with sample IDs, the function expects this
column to be sample_id but the user can override this if they know the name for this column.}
}
\value{
A data frame with statistical scores for the selected samples/subtypes.
}
\description{
Run a Mann-Whitney and general linear regression on a set of
prediction scores.
}
\details{
This function takes a data frame with prediction data \code{these_predictions} and executes
statistical tests to find significant prediction scores based on a categorical variable (i.e response,
tumour grade, etc.) from a provided metadata table. Currently, the function expects the incoming
data to be the score output from \code{\link[=lundtax_predict_sub]{lundtax_predict_sub()}}, together with metadata
information of interest (e.g two level categorical) and subtype classification information. The user
have the option to point the function to the categorical variable with \code{cat_variable}. The return
can be further subset by subtype by using the \code{this_subtype} variable, should be one of the valid
subtypes within the specified class.
}
\examples{
#load packages
library(dplyr, stats)

#get prediction calls
sjodahl_predicted = lundtax_predict_sub(this_data = sjodahl_2017, 
                                        impute = TRUE)

#run general linear models
sjodahl_glm = get_glm_scores(these_predictions = sjodahl_predicted,
                             these_samples_metadata = sjodahl_2017_meta,
                             subtype_class = "5_class",
                             this_subtype = "Uro",
                             cat_variable = "adj_chemo")

}
