% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sample_metrics.R
\name{get_sample_metrics}
\alias{get_sample_metrics}
\title{Get Sample Metrics}
\usage{
get_sample_metrics(
  this_metadata = NULL,
  this_metadata_variable = NULL,
  these_predictions = NULL,
  factor_level = NULL,
  subtype_class = "5_class"
)
}
\arguments{
\item{this_metadata}{A data frame containing metadata with a \code{sample_id} column.}

\item{this_metadata_variable}{A string specifying the column name in the metadata to be used for progression events.}

\item{these_predictions}{A named vector of predictions with sample IDs as names.}

\item{factor_level}{The level of the factor in \code{this_metadata_variable} to summarize (default is 1).}

\item{subtype_class}{The classification system, default is 5 class.}
}
\value{
A data frame with the number of samples and progression events for each subtype.
}
\description{
This function tables a metadata column based on subtype classification.
}
\details{
The function performs the following steps:
\itemize{
\item Ensures the sample IDs are present in both metadata and predictions.
\item Checks if the desired metadata column is valid and is a factor.
\item Filters metadata and predictions to include only common samples.
\item Combines metadata and predictions into a single data frame.
\item Counts the number of samples and progression events for each subtype.
}
}
\examples{
my_predictions = lundtax_predict_sub(this_data = sjodahl_2017, 
                                     gene_id = "hgnc_symbol", 
                                     impute = TRUE, 
                                     adjust = TRUE)

get_sample_metrics(these_predictions = my_predictions, 
                   this_metadata = sjodahl_2017_meta, 
                   this_metadata_variable = "surv_os_event", 
                   factor_level = 1)

}
