% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/int_calc_score.R
\name{int_calc_score}
\alias{int_calc_score}
\title{Calculate Scores.}
\usage{
int_calc_score(
  this_data = NULL,
  variable = NULL,
  logTransform = FALSE,
  gene_id = "hgnc_symbol",
  adjust = TRUE,
  adj_factor = 5.1431
)
}
\arguments{
\item{this_data}{Required parameter. Data frame or matrix with expression values.}

\item{variable}{Required parameter. Input should be one of the following; immune,
score141up, proliferation, or progression.}

\item{gene_id}{Specify the type of gene identifier used in \code{this_data}.
Accepted values are; hgnc_symbol (default) or ensembl_gene_id.}

\item{adjust}{Boolean parameter. If TRUE, the function will proceed with
adjusting the scores based on stable genes. If FALSE (default), no adjustment
will be made and the original score values will be retained.}

\item{adj_factor}{Only applicable if adjust is set to TRUE. Allows users to
apply a proportional adjustment to the normalized scores, enabling finer
control over the final output values. After dividing each score by the mean
expression of stable genes, the result is multiplied by this factor.
Default is 5.1431}

\item{log_transform}{Boolean parameter. If TRUE, the function log transforms
the incoming expression values. Default is FALSE.}
}
\value{
A data frame with scores for the selected variable.
}
\description{
Calculate immune and infiltration scores.
}
\details{
Internal function called by \code{score_lundtax}.
Not meant for out of package use. Takes a data frame of matrix with expression
values and calculates scores based on gene expression.
}
\examples{
#load packages
library(dplyr, multiclasspairs)

#calculate immune scores from hgnc symbols 
immune_scores = int_calc_score(this_data = sjodahl_2017, 
                                variable = "immune", 
                                gene_id = "hgnc_symbol")

#calculate progression score from hgnc symbols 
141_scores = int_calc_score(this_data = sjodahl_2017, 
                            variable = "score141up", 
                            gene_id = "hgnc_symbol")
}
