% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/int_check_ties.R
\name{int_check_ties}
\alias{int_check_ties}
\title{Check Ties.}
\usage{
int_check_ties(first, last)
}
\arguments{
\item{first}{Predictions when setting ties.method = "first"}

\item{last}{Predictions when setting ties.method = "last"}
}
\value{
Nothing.
}
\description{
Check score ties between predictions scores.
}
\details{
Internal function called by \code{\link[=lundtax_calc_sigscore]{lundtax_calc_sigscore()}}.
Not meant for out of package use. Checks score ties in the prediction scores and prints a message
indicating the sample where the tie occurred, the two subtypes with the tied scores and the
subtype that is reported in the output object.
}
