% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/int_get_prop_scores.R
\name{int_get_prop_scores}
\alias{int_get_prop_scores}
\title{Get Proportional Scores}
\usage{
int_get_prop_scores(these_predictions = NULL)
}
\arguments{
\item{these_predictions}{A list containing a data frame named \code{scores} with prediction scores.
The data frame should have row names as sample IDs and columns corresponding to different cell types and scores.}
}
\value{
A data frame with the original scores, normalized scores, and the calculated
immune/stroma ratio. The row names are the sample IDs.
}
\description{
Internal function called by \code{plot_hm_scores} when \code{proportional_scores = TRUE}.
Not meant for out-of-package usage.
}
\details{
This function normalizes specific prediction scores so that each value is a fraction of
the total score for that row. It also calculates the immune/stroma ratio.
}
\examples{
#load packages
library(dplyr)

#get prediction calls
sjodahl_predicted = lundtax_predict_sub(this_data = sjodahl_2017, 
                                        impute = TRUE)
                                        
#transform data into proportional signature scores
prop_scores = int_get_prop_scores(these_predictions = sjodahl_predicted)

}
