% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{Lund2017}
\alias{Lund2017}
\title{Gene expression data derived from the Sjödahl et. al. (2017) cohort}
\format{
An object of class \code{data.frame} with 15697 rows and 301 columns.
}
\usage{
Lund2017
}
\description{
Matrix of RMA normalized gene expression values for 15697 genes with
HGNC symbols and 301 samples
}
\references{
\url{https://onlinelibrary.wiley.com/doi/full/10.1002/path.4886}
}
\keyword{datasets}
