
\docType{data}
\name{lund_colors}
\alias{lund_colors}
\title{Colors used for Lund Taxonomy subtypes}
\format{
An object of class \code{list} of length 4.
}
\usage{
cols
}
\description{
Colors used for Lund Taxonomy subtypes
List including hexadecimal color codes for LundTax subtypes, stage and datasets
}
\keyword{datasets}
