% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{package}
\name{LundTax2023Classifier-package}
\alias{LundTax2023Classifier}
\alias{LundTax2023Classifier-package}
\title{LundTax2023}
\description{
This packages implements a Random Forest rule-based single-sample predictor that classifies
transcriptomic samples into the 5 (or 7, including subclasses) Lund Taxonomy molecular subtypes.
The final classifier is composed of two separate predictors applied sequentially: first a sample
is classified as one of the 5 main classes (Uro, GU, BaSq, Mes or ScNE), and then samples classified as
Uro are subclassified into UroA, UroB or UroC by a second predictor
}
\author{
\strong{Maintainer}: Elena Aramendía \email{elena.aramendia_cotillas@med.lu.se}

Authors:
\itemize{
  \item Pontus Eriksson \email{pontus.eriksson@med.lu.se}
  \item Adam Mattsson \email{adam.mattsson@med.lu.se}
}

}
