% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lundtax_predict_sub.R
\name{lundtax_predict_sub}
\alias{lundtax_predict_sub}
\title{Lund Taxonomy Predict Subtype}
\usage{
lundtax_predict_sub(
  this_data = NULL,
  gene_id = "hgnc_symbol",
  threshold_progression = 0.58,
  log_transform = TRUE,
  adjust = FALSE,
  adj_factor = 5.1431,
  impute = FALSE,
  impute_reject = 0.67,
  impute_kNN = 5,
  subtype_only = FALSE,
  include_data = FALSE,
  include_pred_scores = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{this_data}{Required parameter. Data frame or matrix with expression values.}

\item{gene_id}{Specify the type of gene identifier used in \code{this_data}. Accepted values are;
hgnc_symbol (default) or ensembl_gene_id.}

\item{threshold_progression}{Threshold to flag a sample as high risk of progression, default is 0.58.}

\item{log_transform}{Boolean parameter. If TRUE, the function log transforms the incoming expression
values. Default is TRUE.}

\item{adjust}{Boolean parameter. If TRUE, the function will proceed with adjusting the scores based
on stable genes. If FALSE (default), no adjustment will be made and the original score values will be retained.}

\item{adj_factor}{Only applicable if adjust is set to TRUE. Allows users to apply a proportional
adjustment to the normalized scores, enabling finer control over the final output values.
After dividing each score by the mean expression of stable genes, the result is multiplied by this factor.}

\item{impute}{From \code{\link[multiclassPairs:predict_RF]{multiclassPairs::predict_RF()}}. Boolean. To determine if missed genes and NA
values should be imputed or not. The non missed rules will be used to determine the closest samples
in the training binary matrix (i.e. which is stored in the classifier object). For each sample,
the mode value for nearest samples in the training data will be assigned to the missed rules. Default is FALSE.}

\item{impute_reject}{From \code{\link[multiclassPairs:predict_RF]{multiclassPairs::predict_RF()}}. A number between 0 and 1 indicating the
threshold of the missed rules in the sample. Based on this threshold the sample will be rejected
(i.e. skipped if higher than the impute_reject threshold) and the missed rules will not be imputed
in this sample. Default is 0.67. NOTE, The results object will not have any results for this sample.}

\item{impute_kNN}{From \code{\link[multiclassPairs:predict_RF]{multiclassPairs::predict_RF()}}. Integer determines the number of the nearest
samples in the training data to be used in the imputation. Default is 5. It is not recommended to
use large number (i.e. >10).}

\item{subtype_only}{Boolean parameter. Set to TRUE to return subtypes and nothing else. Default is FALSE.}

\item{include_data}{Boolean parameter. Set to TRUE to include data in output, FALSE is default.}

\item{include_pred_scores}{Boolean parameter. Set to TRUE (default) to include
prediction scores for each sample and class in output.}

\item{verbose}{A logical value indicating whether processing messages will be
printed or not. Default is TRUE.}
}
\value{
Returns a list object including: Data (optional, not included by default),
Prediction scores for all classes (optional, included by default), Predicted LundTax
class for 7-class system, Predicted LundTax class for 5-class system, as well a data frame with
missing genes information.
}
\description{
Predict Lund Taxonomy subtypes based on rule-based Random Forest classifiers.
}
\details{
This function uses 2 classifiers to classify the samples: 5-class
classifier first  classifies samples into Uro, GU, BaSq, Mes or ScNE.
Samples classified as Uro receive a second classification as UroA, B or C by
the second classifier. This function internally calls
\code{\link[=lundtax_calc_sigscore]{lundtax_calc_sigscore()}} for retrieving signature scores.
}
\examples{
sjodahl_predicted = lundtax_predict_sub(this_data = sjodahl_2017, 
                                        impute = TRUE)

}
