% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{LundTax_RF_5c}
\alias{LundTax_RF_5c}
\title{Classifier as a 'rule_based_RandomForest' object. Predicts samples as one of
the 5 main Lund Taxonomy molecular subtypes, Uro, GU, BaSq, Mes, or ScNE
Object includes the final RF classifier, the used genes and rules in the final model,
the Boruta results, and the training matrix. The training matrix is a binary matrix
containing the rule values for the training data and it is used for
imputation purposes during the prediction if values are missing in the sample.
This object was generated using the multiclassPairs package}
\format{
An object of class \code{rule_based_RandomForest} of length 1.
}
\usage{
LundTax_RF_5c
}
\description{
Classifier as a 'rule_based_RandomForest' object. Predicts samples as one of
the 5 main Lund Taxonomy molecular subtypes, Uro, GU, BaSq, Mes, or ScNE
Object includes the final RF classifier, the used genes and rules in the final model,
the Boruta results, and the training matrix. The training matrix is a binary matrix
containing the rule values for the training data and it is used for
imputation purposes during the prediction if values are missing in the sample.
This object was generated using the multiclassPairs package
}
\references{
\url{https://github.com/NourMarzouka/multiclassPairs}
}
\keyword{datasets}
