% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/int_merge_suburo_matrix.R
\name{merge_suburo_matrix}
\alias{merge_suburo_matrix}
\title{Merge subUro Matrix.}
\usage{
merge_suburo_matrix(score_matrix1, score_matrix2, row.names = list(NULL, NULL))
}
\arguments{
\item{score_matrix1}{Prediction score matrix from the 7 class classifier.}

\item{score_matrix2}{Prediction score matrix from the 5 class classifier.}

\item{row.names}{Rownames (sample names) for both matrices.}
}
\value{
Merged matrix including scores for 8 classes (Uro, UroA, UroB, UroC,
GU, BaSq, Mes, ScNE)
}
\description{
Merge prediction score matrices from two classifiers.
}
\details{
Internal function called by \code{predict_lundtax}. Not meant for out of
package use. This function merges the prediction score matrices from the
5-class and 7-class (UroA,UroB,UroC) classifiers into 1 unique score matrix.
}
