% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_hm_signatures.R
\name{plot_hm_signatures}
\alias{plot_hm_signatures}
\title{Plot Heatmap Signatures.}
\usage{
plot_hm_signatures(
  these_predictions = NULL,
  this_data = NULL,
  gene_id = "hgnc_symbols",
  subtype_annotation = "5_class",
  this_sample_order = NULL,
  norm = TRUE,
  plot_scores = TRUE,
  show_ann_legend = FALSE,
  show_hm_legend = FALSE,
  ann_height = 8,
  plot_title = "My Plot",
  plot_width = 14,
  plot_height = 11,
  plot_font_size = 10,
  plot_font_row_size = 7,
  out_path = NULL,
  out_format = "png"
)
}
\arguments{
\item{these_predictions}{Required parameter, should be the output from
\code{\link[=lundtax_predict_sub]{lundtax_predict_sub()}}.}

\item{this_data}{Expression data used for predictions.
Required if the output from \code{\link[=lundtax_predict_sub]{lundtax_predict_sub()}} is run with
include_data = FALSE (default).}

\item{gene_id}{Specify the type of gene identifier used in \code{this_data}. Accepted values are;
hgnc_symbol (default) or ensembl_gene_id.}

\item{subtype_annotation}{Can be one of the following; "5 _class" (default) or "7_class"
annotation.}

\item{this_sample_order}{Optional, set sample order. Default, samples are split by subtype, and order within
each subtype. By default, samples are order by late/early cell cycle ratio (low to high).}

\item{norm}{Boolean parameter. Set to TRUE (default) to normalize the data into Z-scaled values.}

\item{plot_scores}{Boolean parameter. Set to TRUE (default) to plot prediction scores for each
class.}

\item{show_ann_legend}{Boolean parameter, set to TRUE to show annotation legend (Lund classes).
Default is FALSE.}

\item{show_hm_legend}{Boolean parameter, set to TRUE to show heatmap legend, default is FALSE.}

\item{ann_height}{Plotting parameter, optional. Annotation height in cm. Default = 8.}

\item{plot_title}{Plotting parameter. The title for the generated heatmap. Deafult is "My Plot".}

\item{plot_width}{This parameter controls the width in inches. Default is 14 (4200 pixels at 300
PPI).}

\item{plot_height}{This parameter controls the height in pixels. Default is 10 (3000 pixels at
300 PPI)}

\item{plot_font_size}{Optional parameter to control the size of the font in the generated
heatmap. Note, the title of the plot will always be twice that of the set value here
(default = 10).}

\item{plot_font_row_size}{Optional parameter to control the size of the font in the generated
heatmap. Note, the title of the plot will always be twice that of the set value here
(default = 8).}

\item{out_path}{Optional, set path to export plot.}

\item{out_format}{Required parameter if \code{out_path} is specified. Can be "png" (default) or "pdf".
The user can control the dimensions with \code{plot_width} and \code{plot_height}.}
}
\value{
Draws heatmap and silently returns the sample order.
}
\description{
Plot heatmap for classification results.
}
\details{
This function plots a heatmap including genes and signatures of
interest, with prediction results and scores on top.
}
\examples{
\dontrun{ 
#example 1 including data in results object
#run predictor on the bundled expression data
sjodahl_predicted = lundtax_predict_sub(this_data = sjodahl_2017,
                                        include_data = TRUE, 
                                        impute = TRUE)



#example 2 - 5 class annotation and Without prediction scores
plot_hm_signatures(these_predictions = sjodahl_predicted, 
                   subtype_annotation = "5_class",
                   ann_height = 0.5,
                   plot_scores = FALSE)
}

}
