% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ratio_forest.R
\name{plot_ratio_forest}
\alias{plot_ratio_forest}
\title{Forest Plot}
\usage{
plot_ratio_forest(
  these_predictions = NULL,
  these_samples_metadata = NULL,
  this_data = NULL,
  stat_plot,
  all_subs = FALSE,
  this_subtype = NULL,
  subtype_class = "5_class",
  scale = NULL,
  bin_scores = FALSE,
  n_bins = 10,
  surv_time = NULL,
  surv_event = NULL,
  categorical_factor = NULL,
  predictor_columns = NULL,
  exclude_columns = NULL,
  significant_p = 0.05,
  sample_id_col = NULL,
  row_to_col = FALSE,
  out_path = NULL,
  out_format = "png",
  file_name = "my_plot",
  plot_title = "My Plot",
  plot_subtitle = "",
  plot_caption = "",
  plot_width = 8,
  plot_height = 8,
  plot_order = NULL,
  plot_arrange = TRUE,
  return_data = FALSE
)
}
\arguments{
\item{these_predictions}{Required parameter if \code{this_data} is not provided. Should be output from
\code{\link[=lundtax_predict_sub]{lundtax_predict_sub()}}.}

\item{these_samples_metadata}{Required parameter if \code{this_data}is not provided. Metadata associated
with he prediction output. Also possible for the user to provide a metadata subset with samples
of interest, the return will be restricted to the samples within the specified group}

\item{this_data}{Required parameter if \code{these_predictions} and \code{these_samples_metadata} is not
provided. Should be the output from \code{get_survival} or \code{get_glm}, depending on the plot type set with \code{stat_plot}.}

\item{stat_plot}{Required parameter. Sets the plot type, for hazard ratio, set to \code{hazard_ratio}.
For odds ratio, set to \code{odds_ratio}.}

\item{this_subtype}{Optional Specify subtype of interest. Leave as NULL to not separate statistics on subtype.}

\item{subtype_class}{Can be one of the following; 5_class or 7_class. Default is 5_class.}

\item{scale}{Optional parameter. A numeric value to scale the numeric scores. If provided, all
numeric scores will be multiplied by this value.}

\item{bin_scores}{Boolean parameter. Set to TRUE to bin the numeric scores into discrete bins. Default is FALSE.}

\item{n_bins}{Optional parameter. The number of bins to use when binning numeric scores. Default is 10.}

\item{surv_time}{Required parameter if \code{stat_plot} is set to \code{hazard_ratio}, should be the name
of the column in the metadata with survival time. Should be of value numeric.}

\item{surv_event}{Required parameter if \code{stat_plot} is set to \code{hazard_ratio}, should be the name
of the column in the metadata with survival event. Should be of value factor, with two levels.}

\item{categorical_factor}{Required parameter if \code{this_glm}is not provided. This should be the categorical
variable that is intended for testing. In addition, this should also be a variable of type factor,
with exactly 2 levels.}

\item{predictor_columns}{Optional, should be a vector with column names, either from the provided
metadata or signature score object, to be tested for. If not provided, the function will subset
data to the signature scores returned with \code{lundtax_predict_sub}.}

\item{exclude_columns}{Optional argument, specify columns you wish to exclude from the standard
predictor columns. Note, this parameter is only validated if predictor_columns is NULL (default).}

\item{significant_p}{Numeric parameter for flagging significant p values. Default is 0.05.}

\item{sample_id_col}{Optional parameter. Allows the user to manually specify the name of a column with sample ID.}

\item{row_to_col}{Optional parameter, set to TRUE to convert row names in metadata to a new column
called sample_id. Default is FALSE.}

\item{out_path}{Optional, set path to export plot.}

\item{out_format}{Required parameter if \code{out_path} is specified. Can be "png" (default) or "pdf".
The user can further specify the dimensions of the returned plot with \code{plot_width} and \code{plot_height}.}

\item{file_name}{Optional, if plot is being saved to disk, specify the file name for the file.}

\item{plot_title}{Title for plot.}

\item{plot_subtitle}{Caption for plot.}

\item{plot_width}{This parameter controls the width in inches.  Default is 8 (2400 pixels at 300 PPI).}

\item{plot_height}{This parameter controls the height in inches. Default is 8 (2400 pixels at 300 PPI).}

\item{plot_order}{Optional parameter for setting the order of scores in the returned plot.
Only applies if \code{arrange_plot} is set to TRUE.}

\item{plot_arrange}{Boolean parameter, if set to TRUE the user can specify the order of the score
levels on the y axis. Default is FALSE.}

\item{return_data}{Boolean parameter, set to TRUE and return the formatted data used for plotting.
Default is FALSE}
}
\value{
A forest plot as grub object.
}
\description{
Construct a forest plot using Cox model or generalized linnear model for a set of
signature scores.
}
\details{
Depending on the user defines \code{stat_plot}, this function internally calls \code{get_survival}
or \code{get_glm}, which internally calls \code{int_prediction_wrangler}. The user can also provide the
plotting data with \code{this_data} directly. If \code{plot_data} is set to \code{odds_ratio} the function will
visualize odds ratio for one selected categorical variable. The user sets this variable with \code{categorical_factor}.
This should be a valid column name from either the metadata provided with \code{these_samples_metadata}
or from the signature score data subset. The user can then set what numerical columns that are to
be tested for association. This is controlled with \code{predictor_columns}. The input for this
parameter should match column names corresponding to numeric variables in the metadata or signature
score data subset. Alternatively, the function can also visualize hazard ratios calculated with
\code{get_surv}. To do so, set \code{plot_data = "hazard_ratio"}. If so, the user also needs to point the
function to the corresponding survival columns in the provided metadata. This is done with \code{surv_time}
and \code{surv_event}. To return a plot for just one subtype, specify the subtype and class with
\code{this_subtype} and \code{subtype_class}. To return a plot for more than one subtype, it is recommended
to provide the plot data directly with \code{this_data}.
}
\examples{
#load pacakges
library(dplyr, ggplot2)

#get prediction calls
sjodahl_predicted = lundtax_predict_sub(this_data = sjodahl_2017, 
                                        impute = TRUE)

#hazard ratio                                      
plot_ratio_forest(these_predictions = sjodahl_predicted,
                  these_samples_metadata = sjodahl_2017_meta,
                  stat_plot = "hazard_ratio",
                  plot_subtitle = "n Samples: 267",
                  plot_title = "All 7 class samples",
                  subtype_class = "7_class", 
                  this_subtype = NULL,
                  surv_time = "surv_css_time",
                  surv_event = "surv_css_event")

#odds ratio
plot_ratio_forest(these_predictions = sjodahl_predicted,
                  these_samples_metadata = sjodahl_2017_meta,
                  stat_plot = "odds_ratio",
                  plot_title = "Uro Samples - Adjuvant Chemo",
                  plot_subtitle = "n Samples: 121",
                  subtype_class = "5_class", 
                  this_subtype = "Uro",
                  predictor_columns = c("progression_score", 
                                        "proliferation_score", 
                                        "monocytic_lineage"),
                  categorical_factor = "adj_chemo")

}
