% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_subscore_violin.R
\name{plot_subscore_violin}
\alias{plot_subscore_violin}
\title{Violin Plot Subtype Scores}
\usage{
plot_subscore_violin(
  these_predictions,
  this_subtype,
  plot_title = NULL,
  out_path = NULL,
  out_format = "png",
  plot_width = 4,
  plot_height = 4,
  plot_adjust = 2,
  plot_scale = "width",
  plot_trim = FALSE,
  return_data = FALSE
)
}
\arguments{
\item{these_predictions}{Required parameter, should be the output from
\code{\link[=lundtax_predict_sub]{lundtax_predict_sub()}}.}

\item{this_subtype}{Required parameter. Should be one of the set subtype classes from the
LundTax nomenclature.}

\item{plot_title}{Required parameter, if \code{out_path} is specified. plot title, will also be pasted to
the exported file.}

\item{out_path}{Optional, set path to export plot.}

\item{out_format}{Required parameter if \code{out_path} is specified. Can be "png" (default) or "pdf".
The user can further specify the dimensions of the returned plot with \code{plot_width} and \code{plot_height}.}

\item{plot_width}{This parameter controls the width in inches.
Default is 4 (1200 pixels at 300 PPI).}

\item{plot_height}{This parameter controls the height in inches.
Default is 4 (1200 pixels at 300 PPI).}

\item{plot_adjust}{A multiplicate bandwidth adjustment. This makes it possible to adjust the
bandwidth while still using the a bandwidth estimator. For example, adjust = 1/2 means use half
of the default bandwidth.}

\item{plot_scale}{if "area" (default), all violins have the same area (before trimming the tails).
If "count", areas are scaled proportionally to the number of observations. If "width", all
violins have the same maximum width.}

\item{plot_trim}{If TRUE (default), trim the tails of the violins to the range of the data. If
FALSE, don't trim the tails.}

\item{return_data}{Set to TRUE to return tidy data used by the plotting function. Default is FALSE.}
}
\value{
Nothing.
}
\description{
Visualize subtype prediction score for a set subtype in a violin plot.
}
\details{
Take the output from \code{lundtax_predict_sub} and return a violin plot where each violin
represent the distribution of the, for that class, subtype prediction score. Set the subtype of
desire with \code{this_subtype}. The subtype can be one of the subtypes included in the LundTax subtype
classification nomenclature. Parameters starting with the plot prefix are available for the user
to gain more control of the returned plot. The user can also return the wrangled data used in the
plot by setting \code{return_data = TRUE}.
}
\examples{
my_predictions = lundtax_predict_sub(these_predictions = sjodahl_2017, 
                                     gene_id = "hgnc_symbol", 
                                     impute = TRUE, 
                                     adjust = TRUE)
                                     
uro_scores = plot_subscore_violin(these_predictions = my_predictions, 
                                  this_subtype = "Uro")
                                  
}
