% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_lundtax.R
\name{predict_lundtax}
\alias{predict_lundtax}
\title{Predict Lund Taxonomy subtypes based on rule-based Random Forest classifiers}
\usage{
predict_lundtax(
  data,
  subtype_only = FALSE,
  include_data = FALSE,
  include_pred_scores = TRUE,
  gene_id = c("hgnc_symbol", "ensembl_gene_id")[1],
  scoring_method = c("ratio", "singscore")[1],
  logTransform = TRUE,
  adjust = TRUE,
  adj_factor = 5.1431,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{data}{matrix, data frame or multiclassPairs_object of gene expression values}

\item{subtype_only}{}

\item{include_data}{include data in output (disabled by default)}

\item{include_pred_scores}{include prediction scores for each sample and class in output (default)}

\item{gene_id}{specify the type of gene identifier used in the data:
\itemize{
\item "hgnc_symbol" for HUGO gene symbols
\item "ensembl_gene_id" for Ensembl gene IDs
\item "entrezgene" for Entrez IDs
Default value is hgnc_symbol
}}

\item{logTransform}{h}

\item{adjust}{hh}

\item{adj_factor}{default is 5.1431}

\item{...}{Additional parameters to be passed to the predict_RF function. If genes are missing in the data, include impute = TRUE here}
}
\value{
Returns a list object including:
\itemize{
\item Data (optional, not included by default)
\item Prediction scores for all classes (optional, included by default)
\item Predicted LundTax class for 7-class system
\item Predicted LundTax class for 5-class system
}
}
\description{
Predict Lund Taxonomy subtypes based on rule-based Random Forest classifiers
}
\details{
This function uses 2 classifiers to classify the samples: 5-class classifier first  classifies samples into Uro, GU, BaSq, Mes or ScNE.
Samples classified as Uro receive a second classification as UroA, B or C by the second classifier
}
\examples{
# Include data in result
results_data <- predict_lundtax(Lund2017,
                                impute = TRUE,
                                include_data = TRUE)

}
