% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_functions.R
\name{print.rule_based_RandomForest}
\alias{print.rule_based_RandomForest}
\title{Print method for the RF classifier}
\usage{
\method{print}{rule_based_RandomForest}(x, ...)
}
\description{
Print method for the RF classifier
}
