% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_functions.R
\name{ratio_score}
\alias{ratio_score}
\title{Calculate proliferation and progression scores}
\usage{
ratio_score(
  Data,
  variable = c("proliferation", "progression", NULL)[3],
  logTransform = FALSE,
  gene_id = c("ensembl_gene_id", "hgnc_symbol")[1],
  method = c("ratio", "singscore")[1]
)
}
\arguments{
\item{Data}{matrix or data frame of gene expression values}

\item{variable}{score to calculate: proliferation or progression}

\item{logTransform}{if TRUE, log transform data. Set to FALSE is data is already log transformed}

\item{gene_id}{specify the type of gene identifier used in the data:
\itemize{
\item "hgnc_symbol" for HUGO gene symbols
\item "ensembl_gene_id" for Ensembl gene IDs
\item "entrezgene" for Entrez IDs
Default value is hgnc_symbol
}}
}
\value{
Returns values of the calculated score for each sample.
}
\description{
Calculate proliferation and progression scores
}
\examples{
# Calculate proliferation score, hgnc symbols 
results <- ratio_score(Lund2017, variable = "proliferation", gene_id = "hgnc_symbol")

# Calculate progression score, hgnc symbols 
results <- ratio_score(Lund2017, variable = "progression", gene_id = "hgnc_symbol")
                                         impute = TRUE)

}
