% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_lundtax.R
\name{score_lundtax}
\alias{score_lundtax}
\title{Lund Taxonomy Scores.}
\usage{
score_lundtax(
  this_data = NULL,
  gene_id = "hgnc_symbol",
  threshold_progression = 0.58,
  logTransform = TRUE,
  adjust = FALSE,
  adj_factor = 5.1431,
  impute = FALSE,
  impute_reject = 0.67,
  impute_kNN = 5,
  verbose = TRUE
)
}
\arguments{
\item{this_data}{Required parameter. Data frame or matrix with expression values.}

\item{gene_id}{Specify the type of gene identifier used in \code{this_data}. Accepted values are;
hgnc_symbol (default) or ensembl_gene_id.}

\item{threshold_progression}{Threshold to flag a sample as high risk of progression, default is 0.58.}

\item{adjust}{Boolean parameter. If TRUE, the function will proceed with adjusting the scores based
on stable genes. If FALSE (default), no adjustment will be made and the original score values will be retained.}

\item{adj_factor}{Only applicable if adjust is set to TRUE. Allows users to apply a proportional
adjustment to the normalized scores, enabling finer control over the final output values.
After dividing each score by the mean expression of stable genes, the result is multiplied by this factor.}

\item{impute}{From \code{\link[multiclassPairs:predict_RF]{multiclassPairs::predict_RF()}}. Boolean. To determine if missed genes and NA
values should be imputed or not. The non missed rules will be used to determine the closest samples
in the training binary matrix (i.e. which is stored in the classifier object). For each sample,
the mode value for nearest samples in the training data will be assigned to the missed rules. Default is FALSE.}

\item{impute_kNN}{From \code{\link[multiclassPairs:predict_RF]{multiclassPairs::predict_RF()}}. Integer determines the number of the nearest
samples in the training data to be used in the imputation. Default is 5. It is not recommended to
use large number (i.e. >10).}

\item{log_transform}{Boolean parameter. If TRUE (default), the function log transforms the incoming expression
values.}
}
\value{
A data frame with scores for the selected variable.
}
\description{
Wrapper function for calculating Lund Taxonomy scores.
}
\details{
This function internally calls a set of non-exported scoring functions starting with the
int_ prefix. See individual function documentation for more information on the individual function calls.
}
\examples{
#load packages
library(dplyr, multiclassPairs)

#get scores for bundled data
my_scores = score_lundtax(this_data = sjodahl_2017, impute = TRUE)
                            
}
