% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{signatures}
\alias{signatures}
\title{Signature table}
\format{
An object of class \code{data.frame} with 475 rows and 2 columns.
}
\usage{
signatures
}
\description{
Dataframe containing gene symbols and corresponding signature for
each gene included in the following gene singatures: FGFR3, Keratinization,
early cell cycle, late cycle, Immune141 UP and Stromal141UP
}
\keyword{datasets}
