% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_functions.R
\name{single_score}
\alias{single_score}
\title{Calculate immune, infiltration (141 UP) and prostate scores}
\usage{
single_score(
  Data,
  variable = c("immune", "score141up", "prostate", NULL)[4],
  logTransform = FALSE,
  gene_id = c("ensembl_gene_id", "hgnc_symbol")[1],
  adjust = TRUE,
  adj_factor = 5.1431
)
}
\arguments{
\item{Data}{matrix or data frame of gene expression values}

\item{variable}{score to calculate: proliferation or progression}

\item{logTransform}{if TRUE, log transform data. Set to FALSE is data is already log transformed}

\item{gene_id}{specify the type of gene identifier used in the data:
\itemize{
\item "hgnc_symbol" for HUGO gene symbols
\item "ensembl_gene_id" for Ensembl gene IDs
\item "entrezgene" for Entrez IDs
Default value is hgnc_symbol
}}
}
\value{
Returns values of the calculated score for each sample.
}
\description{
Calculate immune, infiltration (141 UP) and prostate scores
}
\examples{
# Calculate immune score, hgnc symbols 
results <- single_score(Lund2017, variable = "immune", gene_id = "hgnc_symbol")

# Calculate progression score, hgnc symbols 
results <- single_score(Lund2017, variable = "prostate", gene_id = "hgnc_symbol")
                                         impute = TRUE)

}
