% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sjodahl_2017}
\alias{sjodahl_2017}
\title{Sjodahl 2017.}
\format{
A data frame with 15697 rows (gene expressions) and 267 columns (samples).
}
\usage{
data(sjodahl_2017)
}
\description{
Gene expression data derived from the Sjödahl et. al. (2017) cohort. A matrix of RMA normalized
and ComBat adjusted gene expression values for 15697 genes (hgnc format) for 267 samples
}
\details{
A data frame with gene information in different formats.
}
\keyword{datasets}
