% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sjodahl_2017_meta}
\alias{sjodahl_2017_meta}
\title{Sjodahl 2017 Meta.}
\format{
A data frame with 15697 rows (gene expressions) and 267 columns (samples).
}
\usage{
data(sjodahl_2017_meta)
}
\description{
metadata associated with the bundled expresison data.
}
\details{
A data frame with meta data for the bundled samples.

\itemize{
\item sample_id. Unique sample ID.
\item cluster_order. Cluster order.
\item age. Age of patient, years.
\item gender. Patient gender.
\item region_cx. Geographical region of CX.
\item turb_grade. Pathological grading of TURB-specimen. 0 = WHO1999 Grade 2, 1 = WHO1999 Grade 3.
\item turb_stage. Pathological staging of TURB-specimen. 0 = Non muscle-invasive, 1 = Muscle invasive.
\item turb_skiv. Presence of keratinization or squamous differentiation in TURB-specimen. 0 = No presence, 1 = presence.
\item turb_cis. Presence of cis in the TURB-specimen. 0 = No presence, 1 = presence.
\item turb_lvi. Presence of lymphovascular invasion in the TURB-specimen. 0 = No presence, 1 = presence.
\item turb_necros. Presence of necrosis in TURB-specimen. 0 = no, 1 = present, 2 = extensive.
\item turb_apoptos. Presence of apoptosis in the TURB-specimen. 0 = No presence, 1 = presence.
\item variant_histology. Variant histology present 1 = yes, 0 = no.
\item cT. Clinical tumor state
\item cN. Clinical node state
\item hydroneph. Presence of hydronephrose. 0 = No presence, 1 = presence.
\item pTCx. Pathological tumor state at cystectomy.
\item pN. Patological node state at cystectomy.
\item adj_chemo. Adjeuvant chemotherapy. 0 = No presence, 1 = presence.
\item surv_os_event. Survival data, 1 = event, 0 = non event.
\item surv_os_time. Survival data, in months.
\item surv_css_event. Cancer specific survival, 1 = event, 0 = non event.
\item surv_css_time. Cancer specific survival, in months.
\item surv_pfs_event. Progression-free survival, 1 = event, 0 = non event.
\item surv_pfs_time. Progression-free survival, in months.
}
}
\keyword{datasets}
