% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/int_calc_immune_proportions.R
\name{int_calc_immune_proportions}
\alias{int_calc_immune_proportions}
\title{Calculate Immune Proportions.}
\usage{
int_calc_immune_proportions(immune_results)
}
\arguments{
\item{immune_results}{Required parameter. Data frame with immune scores from int_calc_score.}
}
\value{
A data frame with scores as percentages.
}
\description{
Calculate immune score proportions.
}
\details{
Internal function called by \code{in_calc_score} if varaible is set to immune.
}
