% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lund_colors}
\alias{lund_colors}
\title{Lund Colors.}
\format{
A list of 4.
}
\usage{
data(lund_colors)
}
\description{
Standardized colors used for Lund Taxonomy subtypes and datasets.
}
\details{
A list of 4 with color palettes frequently used in this package.

\itemize{
\item lund_colors. Color palette for the LundTax subtypes.
\item lund_colors_transp. Same as \code{lund_colors} but transparent.
\item stage_colors. Color palette for stages.
\item dataset_colors. Color palette for the LundTax datasets.
}
}
\keyword{datasets}
