% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_functions.R
\name{merge_subUro_matrix}
\alias{merge_subUro_matrix}
\title{Merge prediciton score matrices from two classifiers}
\usage{
merge_subUro_matrix(score_matrix1, score_matrix2, row.names = list(NULL, NULL))
}
\arguments{
\item{score_matrix1}{prediction score matrix from the 7 class classifier}

\item{score_matrix2}{prediction score matrix from the 5 class classifier}

\item{row.names}{rownames (sample names) for both matrices}
}
\value{
Merged matrix including scores for 8 classes (Uro, UroA, UroB, UroC, GU, BaSq, Mes, ScNE)
}
\description{
This function merges the prediction score matrices from the 5-class and 7-class (UroA,UroB,UroC) classifiers into 1 unique score matrix
}
