% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ranked_score.R
\name{plot_ranked_score}
\alias{plot_ranked_score}
\title{Plot Ranked Score}
\usage{
plot_ranked_score(
  these_predictions = NULL,
  this_data = NULL,
  this_score = NULL,
  this_subtype = NULL,
  subtype_class = NULL,
  seg_plot = FALSE,
  seg_width = 0.1,
  add_stat = FALSE,
  title = NULL,
  out_path = NULL,
  out_format = "png",
  plot_width = 4,
  plot_height = 4,
  plot_title = NULL,
  return_data = FALSE
)
}
\arguments{
\item{these_predictions}{Required parameter, if no data provided with \code{this_data}.
Predictions returned with \code{\link[=lundtax_predict_sub]{lundtax_predict_sub()}}.}

\item{this_data}{Optional parameter, makes it possible for the user to give the plotting function
their own data, preferably retrieved with this function with return_data = TRUE. If provided,
the plotting function will disregard any other parameters and only draw the plot using the data
provided.}

\item{this_score}{Required parameter, should be one of the numeric columns in the scores data
frame from the list returned with \code{\link[=lundtax_predict_sub]{lundtax_predict_sub()}}.}

\item{this_subtype}{Optional parameter, lets the user subset the return to a specific subtype.
If nopt provided, all subtypes within the selected class will be returned.}

\item{subtype_class}{Required, one of the following; 5_class or 7_class. Needed for coloring the
points based on subtype classification.}

\item{seg_plot}{Boolean parameter, set to TRUE to return a segment plot with ranked scores for the
selected subtype class. Default is FALSE.}

\item{add_stat}{Boolean parameter, set to TRUE to add regression lines, p value for each regression. Default is FALSE.}

\item{out_path}{Optional, set path to export plot.}

\item{out_format}{Required parameter if \code{out_path} is specified. Can be "png" (default) or "pdf".
The user can further specify the dimensions of the returned plot with \code{plot_width} and \code{plot_height}.}

\item{plot_width}{This parameter controls the width in inches. Default is 4 (1200 pixels at 300 PPI).}

\item{plot_height}{This parameter controls the height in inches. Default is 4 (1200 pixels at 300 PPI).}

\item{plot_title}{Required parameter, if \code{out_path} is specified. plot title, will also be pasted to
the exported file.}

\item{return_data}{Boolean parameter, set to TRUE and return the formatted data used for
plotting. Default is FALSE}
}
\value{
Nothing.
}
\description{
Return a point plot with ranked scores for a set variable.
}
\details{
This function takes predictions returned with \code{\link[=lundtax_predict_sub]{lundtax_predict_sub()}},
together with a score variable, subtype class and returns a point plot (grob) with ranked scores
along the x axis and scores along the y axis. The returned plot will also color fill the points
based on subtype classification.
}
\examples{
sjodahl_2017_predicted = lundtax_predict_sub(this_data = sjodahl_2017, 
                                             impute = TRUE)

plot_ranked_score(these_predictions = sjodahl_2017_predicted, 
                  this_score = "proliferation_score", 
                  subtype_class = "7_class", 
                  plot_title = "Sjodahl 2017")

}
