% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_signatures.R
\name{plot_signatures}
\alias{plot_signatures}
\title{Plot Signature Results.}
\usage{
plot_signatures(
  these_predictions = NULL,
  this_data = NULL,
  gene_id = "hgnc_symbols",
  annotation = "5 classes",
  norm = TRUE,
  plot_scores = TRUE,
  show_ann_legend = FALSE,
  show_hm_legend = FALSE,
  set_order = NULL,
  ann_height = 6,
  font.size = 8,
  title = "My Plot"
)
}
\arguments{
\item{these_predictions}{Required parameter, should be the output from \code{predict_lundtax}.}

\item{this_data}{Expression data used for predictions.
Required if the output from \code{predict_lundtax} is run with include_data = FALSE (default).}

\item{gene_id}{Specify the type of gene identifier used in \code{this_data}. Accepted values are;
hgnc_symbol (default) or ensembl_gene_id.}

\item{annotation}{Can be one of the following; "5 class" (default) or "7 class" annotation.}

\item{norm}{Boolean parameter. Set to TRUE (default) to normalize the data into Z-scaled values.}

\item{plot_scores}{Boolean parameter. Set to TRUE (default) to plot prediction scores for each class.}

\item{show_ann_legend}{Boolean parameter, set to TRUE to show annotation legend (Lund classes). Default is FALSE.}

\item{show_hm_legend}{Boolean parameter, set to TRUE to show heatmap legend, default is FALSE.}

\item{set_order}{Optional parameter. Set sample order. Samples are split by subtype, and order
within each subtype. By default, samples are order by late/early cell cycle ratio (low to high).}

\item{font.size}{Plotting parameter, optional. font size. Default = 8.}

\item{title}{Plotting parameter. The title for the generated heatmap. Deafult is "My Plot".}

\item{ann_heigh}{Plotting parameter, optional. Annotation height in cm. Default = 6.}
}
\value{
Draws heatmap and silently returns the sample order.
}
\description{
Plot heatmap for classification results.
}
\details{
This function plots a heatmap including genes and signatures of interest, with
prediction results and scores on top.
}
\examples{
#example 1 including data in results object
sjodahl_predicted = predict_lundtax(this_data = sjodahl_2017,
                                    include_data = TRUE)
                          
plot_signatures(these_predictions = sjodahl_predicted)

#example 2 - loading data separately
sjodahl_predicted = predict_lundtax(this_data = sjodahl_2017)

plot_signatures(these_predictions = sjodahl_predicted, 
                this_data = sjodahl_2017)

#example 3 - 5 class annotation and Without prediction scores
plot_signatures(these_predictions = sjodahl_predicted, 
                this_data = sjodahl_2017,
                annotation = "5 classes",
                ann_height = 0.5,
                plot_scores = FALSE)

}
