% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_function.R
\name{plot_signatures}
\alias{plot_signatures}
\title{Plot heatmap for classification results}
\usage{
plot_signatures(
  results_object,
  data = NULL,
  title = "",
  annotation = c("5 classes", "7 classes")[2],
  plot_scores = TRUE,
  show_ann_legend = FALSE,
  show_hm_legend = FALSE,
  set_order = NULL,
  ann_height = 6,
  font.size = 8,
  norm = c("scale", NULL)[1]
)
}
\arguments{
\item{results_object}{description}

\item{data}{results_object}

\item{title}{heatmap title}

\item{annotation}{use 5 or 7 class annotation}

\item{plot_scores}{plot prediction scores for each class}

\item{show_ann_legend}{show annotation legend (Lund classes)}

\item{show_hm_legend}{show heatmap legend}

\item{set_order}{set sample order. Samples are split by subtype, and order within each subtype. By default, samples are order by late/early cell cycle ratio (low to high)}

\item{font.size}{font size (default = 8)}

\item{norm}{normalize the data into Z-scaled values}

\item{ann_heigh}{annotation height in cm (default = 6)}
}
\value{
Draws heatmap and silently returns the sample order
}
\description{
This function plots a heatmap including genes and signatures of interest, with prediction results and scores on top
}
\examples{
# Including data in results object
results <- predict_LundTax2023(Lund2017,
                              include_data = TRUE)
plot_signatures(results)

# Loading data separately
results <- predict_LundTax2023(Lund2017)
plot_signatures(results, data = Lund2017)

# 5 class annotation and Without prediction scores
plot_signatures(results,
               data = Lund2017,
               annotation = "5 classes",
               ann_height = 0.5,
               plot_scores = FALSE)

# Plot and get sample order
sample_order <- plot_signatures(results,
                               data = Lund2017)
print(head(sample_order))

# Save to pdf
pdf("heatmap_example.pdf", width = 15, height = 10)
plot_signatures(results,
               data = Lund2017,
               ann_height = 0.5,
               plot_scores = FALSE)
dev.off()

}
