% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_subscore_box.R
\name{plot_subscore_box}
\alias{plot_subscore_box}
\title{Plot Stacked Barplot Subtype Score}
\usage{
plot_subscore_box(
  these_predictions,
  this_subtype,
  plot_title = NULL,
  out_path = NULL,
  out_format = "png",
  plot_width = 4,
  plot_height = 4,
  return_data = FALSE
)
}
\arguments{
\item{these_predictions}{Required parameter, should be the output from
\code{\link[=lundtax_predict_sub]{lundtax_predict_sub()}}.}

\item{this_subtype}{Required parameter. Should be one of the set subtype classes from the
LundTax nomenclature.}

\item{plot_title}{Required parameter, if \code{out_path} is specified. plot title, will also be pasted to
the exported file.}

\item{out_path}{Optional, set path to export plot.}

\item{out_format}{Required parameter if \code{out_path} is specified. Can be "png" (default) or "pdf".
The user can further specify the dimensions of the returned plot with \code{plot_width} and \code{plot_height}.}

\item{plot_width}{This parameter controls the width in inches.
Default is 4 (1200 pixels at 300 PPI).}

\item{plot_height}{This parameter controls the height in inches.
Default is 4 (1200 pixels at 300 PPI).}

\item{return_data}{Set to TRUE to return tidy data used by the plotting function. Default is FALSE.}
}
\value{
Nothing.
}
\description{
Visualize subtype prediction score within a set subtype in a stacked barplot.
}
\details{
Take the output from \code{lundtax_predict_sub} and return a stacked barplot plot representing
the distribution of the, for that class, subtype prediction score. Set the subtype of
desire with \code{this_subtype}. The subtype can be one of the subtypes included in the LundTax subtype
classification nomenclature.
}
\examples{
#' my_predictions = lundtax_predict_sub(these_predictions = sjodahl_2017, 
                                     gene_id = "hgnc_symbol", 
                                     impute = TRUE, 
                                     adjust = TRUE)
                                     
uro_scores = plot_subscore_box(these_predictions = my_predictions, 
                               this_subtype = "Uro")

}
