% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_functions.R
\name{predict_LundTax2023}
\alias{predict_LundTax2023}
\title{Predict Lund Taxonomy subtypes based on rule-based Random Forest classifiers}
\usage{
predict_LundTax2023(data, include_data = FALSE, include_scores = TRUE, ...)
}
\arguments{
\item{data}{matrix, data frame or multiclassPairs_object of gene expression values}

\item{include_data}{include data in output (disabled by default)}

\item{include_scores}{include prediciton scores for each sample and class in output (default)}

\item{...}{Additional parameters to be passed to the predict_RF function. If genes are missing in the data, include impute = TRUE here}

\item{versions}{description}
}
\value{
Returns a list object including: Data (optional, not included by default), Prediction scores for all classes (optional, included by default), and Predicted LundTax class for both 7 and 5-class systems
}
\description{
Predict Lund Taxonomy subtypes based on rule-based Random Forest classifiers
}
\details{
This function uses 2 classifiers to classify the samples:
\itemize{
\item 5-class classifier classifies samples into Uro, GU, BaSq, Mes or ScNE
\item Samples classified as Uro receive a second classification as UroA, B or C by the second classifier
}
}
\examples{
data(Lund2017)
predict_LundTax2023(Lund2017)
data(Lund2017)
# Include data in result
predict_LundTax2023(Lund2017,
include_data = TRUE)

data(Lund2017)

# Imputation
predict_LundTax2023(Lund2017,
impute = TRUE)

}
