% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{signatures}
\alias{signatures}
\title{Signature table}
\format{
An object of class \code{list} of length 6.
}
\usage{
signatures
}
\description{
Signature table
}
\keyword{datasets}
