% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{signatures}
\alias{signatures}
\title{Signatures}
\format{
A list of 6.
}
\usage{
data(signatures)
}
\description{
Bundles a set of signatures in different molecular processes and the genes associated.
This dataset is needed for calculating the signature scores.
}
\details{
A list of 6 with different signatures and the associated genes.

\itemize{
\item signatures_plot. Genes in hgnc and ensembl format associated with specific signatures.
\item proliferation. Genes in hgnc and ensembl format associated with proliferation.
\item progression. Genes in hgnc and ensembl format associated with progression.
\item prostate. Genes in hgnc and ensembl format associated with prostate cancer.
\item immune. Genes in hgnc and ensembl format associated with immuno process.
\item stable_genes. Genes in hgnc and ensembl format associated with stable genes.
}
}
\keyword{datasets}
